* in 1.3: Changed types used in rrd_file_t from off_t to size_t to prevent: warning: comparison between signed and unsigned * Misc other off_t -> size_t changes to prevent signed warning * in 1.3: Added rrd_dump_opt_r declaration to prevent undef warning * Explicit cast to time_t of check in rrd_fetch to prevent signed warning * Added vasprintf path to sprintf_alloc to prevent excessive mallocing * in 1.3: fixed STRERROR_R_CHAR_P undef warning
patch by Steven Hartland killing multiplay.co.uk




git-svn-id: svn://svn.oetiker.ch/rrdtool/trunk/program@1808 a5681a0c-68f1-0310-ab6d-d61299d08faa

+++ src/rrd_update.c
@@ -1896,7 +1896,7 @@ static int write_to_rras(
                  scratch_idx = CDP_secondary_val,
                  step_subtract = 2) {
 
-            off_t rra_pos_new;
+            size_t rra_pos_new;
 #ifdef DEBUG
             fprintf(stderr, "  -- RRA Preseek %ld\n", rrd_file->pos);
 #endif
